cast GLuint to int64_t for EncodableValue
When building ivi-homescreen with gcc in Yocto, the following error was
caused.

```
/usr/include/c++/9.3.0/variant:1267:8: error: no type named 'type' in 'struct std::enable_if<false, void>'
|  1267 |  using __accepted_type = __to_type<__accepted_index<_Tp>>;
|       |        ^~~~~~~~~~~~~~~
```

In order to solve this error, I added `static_cast` for GLuint in
EncodableValue().

+++ shell/textures/test_egl/texture_test_egl.cc
@@ -59,16 +59,17 @@ flutter::EncodableValue TextureTestEgl::Create(
   return flutter::EncodableValue(flutter::EncodableMap{
       {flutter::EncodableValue("result"), flutter::EncodableValue(0)},
       {flutter::EncodableValue("textureId"),
-       flutter::EncodableValue(textureId)},
-      {flutter::EncodableValue("width"), flutter::EncodableValue(obj->m_width)},
+       flutter::EncodableValue(static_cast<int64_t>(textureId))},
+      {flutter::EncodableValue("width"),
+       flutter::EncodableValue(static_cast<int64_t>(obj->m_width))},
       {flutter::EncodableValue("height"),
-       flutter::EncodableValue(obj->m_height)},
+       flutter::EncodableValue(static_cast<int64_t>(obj->m_height))},
       {flutter::EncodableValue("GL_target"),
-       flutter::EncodableValue(obj->m_target)},
+       flutter::EncodableValue(static_cast<int64_t>(obj->m_target))},
       {flutter::EncodableValue("GL_format"),
-       flutter::EncodableValue(obj->m_format)},
+       flutter::EncodableValue(static_cast<int64_t>(obj->m_format))},
       {flutter::EncodableValue("GL_textureId"),
-       flutter::EncodableValue(textureId)}});
+       flutter::EncodableValue(static_cast<int64_t>(textureId))}});
 }
 
 void TextureTestEgl::Dispose(void* userdata, GLuint name) {
